# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 1.0.0
  description: Maverick Execution Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users
      parameters:
        - in: query
          name: role
          schema:
            type: string
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Create a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: Created user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get a user by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update a user
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete a user
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # BATCHES
...
  /reports/attendance:
    get:
      operationId: getAttendanceReport
      tags: [reports]
      summary: Get attendance report data
      parameters:
        - in: query
          name: batchId
          schema:
            type: integer
        - in: query
          name: startDate
          schema:
            type: string
        - in: query
          name: endDate
          schema:
            type: string
      responses:
        "200":
          description: Attendance report rows
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AttendanceReportRow"
  /reports/assessments:
    get:
      operationId: getAssessmentReport
      tags: [reports]
      summary: Get assessment score report data
      parameters:
        - in: query
          name: batchId
          schema:
            type: integer
      responses:
        "200":
          description: Assessment report rows
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AssessmentReportRow"
  /reports/toppers:
    get:
      operationId: getTopperReport
      tags: [reports]
      summary: Get topper report data
      parameters:
        - in: query
          name: batchId
          schema:
            type: integer
      responses:
        "200":
          description: Topper report rows
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TopperReportRow"
  /reports/consolidated:
    get:
      operationId: getConsolidatedReport
      tags: [reports]
      summary: Get consolidated candidate report
      parameters:
        - in: query
          name: batchId
          schema:
            type: integer
        - in: query
          name: status
          schema:
            type: string
      responses:
        "200":
          description: Consolidated report rows
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ConsolidatedReportRow"

  # ATTENDANCE ALERTS
...
components:
  schemas:
...
    User:
      type: object
      required: [id, name, email, role, isActive, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, coordinator, trainer]
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: ["string", "null"]
          format: date-time
    UserInput:
      type: object
      required: [name, email, password, role]
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
        role:
          type: string
          enum: [admin, coordinator, trainer]
    UserUpdate:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        role:
          type: string
          enum: [admin, coordinator, trainer]
        isActive:
          type: boolean

    # BATCHES
...
    AttendanceReportRow:
      type: object
      required: [candidateName, batchName, date, status]
      properties:
        candidateId:
          type: string
        candidateName:
          type: string
        batchName:
          type: string
        date:
          type: string
        status:
          type: string
        remarks:
          type: ["string", "null"]
    AssessmentReportRow:
      type: object
      required: [candidateName, batchName, assessmentTitle, assessmentType, score, maxScore, percentage]
      properties:
        candidateName:
          type: string
        batchName:
          type: string
        assessmentTitle:
          type: string
        assessmentType:
          type: string
        score:
          type: number
        maxScore:
          type: number
        percentage:
          type: number
        scheduledDate:
          type: string
    TopperReportRow:
      type: object
      required: [rank, candidateName, batchName, totalScore]
      properties:
        rank:
          type: integer
        candidateName:
          type: string
        batchName:
          type: string
        totalScore:
          type: number
        assessmentScore:
          type: ["number", "null"]
        projectScore:
          type: ["number", "null"]
        attendanceScore:
          type: ["number", "null"]
    ConsolidatedReportRow:
      type: object
      required: [candidateName, batchName, status, attendancePct, avgScore]
      properties:
        candidateId:
          type: string
        candidateName:
          type: string
        batchName:
          type: string
        status:
          type: string
        attendancePct:
          type: number
        avgScore:
          type: number
        joinedAt:
          type: ["string", "null"]
...